fix: login form was disabled after 2nd failed login
On submit the `vaadin-login-form` sets `error` to `false` and `disabled` to `true`. Then the `disabled` status is removed only if `error` is set to `true` (via observer `_errorChanged(error)`).

The issue was that Lit property bindings are one way so when login form sets its internal error status to false, our `login-view` still has `error` `true` (after first login failure) and when we try login for the second time then `this.error = result.error;` doesn't do anything because `this.error` is already `true` (in our view), so it doesn't change and it doesn't trigger the propagation of `error` `true` into login form component.

Fixes #21