feat: keep grid highlighting when filter changes
When the filter changes the list of contacts is re-fetch from the backend. The grid gets a new list of grid items and the `selectedItem` can no longer be found in the new list (because the grid does a shallow instance comparison).

This PR fixes the issue when the highlight for the selected grid item is lost if the filter is updated while a contact is selected.