feat: connect the Dashboard view to the state store
This allows optimistic updates to the dashboard view when the list of contacts changes. The dashboard is immediately updated assuming the change succeeds, before the change is processed by the backend.

In order to implement that, the list of contacts is extracted out of the list view's UI store into an entity store (along with companies). Then, a separate UI store is created for the dashboard view, and both the dashboard UI store and the contact list UI store depend on the data in the entity store.