feat: connect the Dashboard view to the state store
This allows optimistic updates to the dashboard view when the list of contacts changes. The dashboard is immediately updated assuming the change succeeds, before the change is processed by the backend.

In order to implement that, the dashboard view is connected to the redux store, and selects the data it needs from 'Contacts' slice.