feat: add a Companies view
Right now it behaves identical to the contact list view (except for showing companies rather than contacts).
The intention is to have a lazy-loaded list of companies to show how to work with lazy data for grids and combo-boxes, and also to show how to work with relational data in the Redux store.