Fix/is user logged in handling (#32)
* fix: show only one login popup at a time

Update the invalid session middleware to show only one login popup at a time. This fixes the issue when the middleware opened a new popup for every failed endpoint call. e.g. opening the Contacts view triggers several endpoint calls (find(), findAllCompanies(), etc) and navigating to that view with an expired session caused several login popups to be created one on top of another.

+ move the invalid session middleware from a 'magic' `connect-client.ts` file into `index.ts` where it is more clear that it's always used.

* fix: update 'isUserLogged' in local storage when session expires

This PR fixes an issues when the 'isUserLogged' flag in the local storage was not cleared when the user session expires. This causes a brief flash of the (empty) contact list view before a login popup is shown, if the app is reloaded after a session expires.

+ add a 30 days expiration on the `isUserLogged` flag in the local storage
+ refactor the `isUserLogged` flag handling logic into a separate `auth.ts` file