feat: enable code splitting / lazy loading view-specific store slices
Use the `@redux-dynostore/core` library to enable dynamic reducer injection in the global Redux store. This enables keeping the view-specific slices of the store separate, and loading them lazily when needed to minimize the initial bundle.

NOTE: the `@redux-dynostore/core` library does not have TypeScript typings, and requires using ``@ts-ignore` in a couple of places.